Optimized Profile of User and updated removing and editing Articles, created controller for Profile and html

+++ src/main/java/com/site/blog/controllers/ArticlesController.java
@@ -1,8 +1,10 @@
 package com.site.blog.controllers;
 
 import com.site.blog.models.Articles;
+import com.site.blog.models.Authorities;
 import com.site.blog.models.Users;
 import com.site.blog.repo.ArticlesRepository;
+import com.site.blog.repo.AuthoritiesRepository;
 import com.site.blog.repo.UsersRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -10,6 +12,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.ui.Model;
 
 import java.util.Date;
+import java.util.List;
 
 import static com.site.blog.StaticMethods.getCurrentUsername;
 
@@ -21,6 +24,9 @@ public class ArticlesController {
 
     @Autowired
     private UsersRepository usersRepository;
+
+    @Autowired
+    private AuthoritiesRepository authoritiesRepository;
 
     @GetMapping("/articles")
     public String articlesMain(Model model) {
@@ -61,7 +67,7 @@ public class ArticlesController {
     public String articlesPostDetails(@PathVariable(value = "id") long id) {
         Articles articles = articlesRepository.findById(id).orElseThrow();
         int views = articles.getViews();
-        articles.setViews(views+1);
+        articles.setViews(views + 1);
         articlesRepository.save(articles);
         return "redirect:/articles/{id}";
     }
@@ -72,8 +78,20 @@ public class ArticlesController {
             return "redirect:/articles";
         }
         Articles article = articlesRepository.findById(id).orElseThrow();
-        model.addAttribute("article", article);
-        return "articles-templates/articles-edit";
+        Users user = usersRepository.findById(getCurrentUsername()).orElseThrow();
+        if (article.getUsers_id().equals(user)) {
+            model.addAttribute("article", article);
+            return "articles-templates/articles-edit";
+        } else {
+            List<Authorities> authoritiesList = authoritiesRepository.findAllByUsername(user);
+            for (Authorities authority : authoritiesList) {
+                if (authority.getAuthority().equals("ROLE_ADMIN")) {
+                    model.addAttribute("article", article);
+                    return "articles-templates/articles-edit";
+                }
+            }
+        }
+        return "redirect:/articles";
     }
 
     @PostMapping("/articles/{id}/edit")
@@ -89,7 +107,18 @@ public class ArticlesController {
 
     @PostMapping("/articles/{id}/remove")
     public String articlesPostDelete(@PathVariable(value = "id") long id) {
-        articlesRepository.deleteById(id);
+        Articles article = articlesRepository.findById(id).orElseThrow();
+        Users user = usersRepository.findById(getCurrentUsername()).orElseThrow();
+        if (article.getUsers_id().equals(user)) {
+            articlesRepository.deleteById(id);
+        } else {
+            List<Authorities> authoritiesList = authoritiesRepository.findAllByUsername(user);
+            for (Authorities authority : authoritiesList) {
+                if (authority.getAuthority().equals("ROLE_ADMIN")) {
+                    articlesRepository.deleteById(id);
+                }
+            }
+        }
         return "redirect:/articles";
     }
 }
